backend/requirements.txt
# Core Framework
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
python-multipart>=0.0.6,<0.1.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Database
sqlalchemy==1.4.53
alembic==1.12.1
databases[postgresql]==0.8.0
asyncpg==0.29.0

# AI/ML - Computer Vision
torch>=2.0.0,<2.2.0
torchvision>=0.15.0,<0.17.0
transformers>=4.35.0,<4.37.0
timm>=0.9.10,<0.10.0
Pillow>=10.0.0,<11.0.0
opencv-python>=4.8.0,<4.9.0

# AI/ML - NLP
openai>=1.3.0,<2.0.0
google-generativeai>=0.3.0,<0.4.0
langchain>=0.0.330,<0.1.0
langchain-community>=0.0.10,<0.1.0

# Vector Database & RAG
faiss-cpu>=1.7.4,<1.8.0
pinecone-client>=2.2.0,<3.0.0
sentence-transformers>=2.2.0,<3.0.0
chromadb>=0.4.15,<0.5.0

# Audio Processing
openai-whisper==20231117
pydub==0.25.1
soundfile==0.12.1
librosa==0.10.1

# Web Scraping & Data
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.4
numpy==1.25.2

# Caching & Session Management
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP & API
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0

# Utilities
python-slugify==8.0.1
email-validator==2.1.0
phonenumbers==8.13.26

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Additional Dependencies
geopy==2.4.0
pytz==2023.3
schedule==1.2.0
celery==5.3.4
flower==2.0.1

# Weather API
pyowm==3.3.0

# Text Processing
nltk==3.8.1
spacy==3.7.2

# File Processing
python-magic==0.4.27
openpyxl==3.1.2
